Prepare conversation list for an archived screen
